Renamed sprite_db to spriteDb
