+++ source/controllers/SpriteController.cpp
@@ -1,6 +1,6 @@
 #include "SpriteController.h"
 #include "core/globals.h"
-#include "data/sprite_db.h"
+#include "data/spriteDb.h"
 
 bool SpriteController::switchSpriteImpl(SpriteType type, int spriteId, SpriteRegister* out)
 {
